% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROBShiny.R
\name{PROBShiny}
\alias{PROBShiny}
\title{Launch 'PROBShiny' Interface}
\usage{
PROBShiny()
}
\value{
Nothing
}
\description{
PROBShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start PROBShiny

The interactive user interface is to provide an easy way for solving basic probability problems based on relative frequency and subjective approach.
}
\examples{
if(interactive()){
library(rmarkdown)
PROBShiny()
}
}
\keyword{PROBShiny}
